+++ flow/flow.vcxproj.filters
@@ -14,6 +14,7 @@
     <ActorCompiler Include="TDMetric.actor.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="crc32c-generated-constants.cpp" />
     <ClCompile Include="DeterministicRandom.cpp" />
     <ClCompile Include="Error.cpp" />
     <ClCompile Include="FastAlloc.cpp" />
